etl/requirements.txt
psycopg2-binary ~= 2.9.9
python-dotenv ~= 1.0.0
pytest~=7.4.3
pytest-mock~=3.12.0
